Move last cache insert into StorageGroupProcessor
